+++ src/main/java/pretzel/dreamketcherbe/domain/member/entity/SocialType.java
@@ -1,0 +1,6 @@
+package pretzel.dreamketcherbe.domain.member.entity;
+
+public enum SocialType {
+    GOOGLE,
+    ;
+}
