--- src/main/java/pretzel/dreamketcherbe/member/entity/SocialType.java
@@ -1,6 +1,0 @@
-package pretzel.dreamketcherbe.member.entity;
-
-public enum SocialType {
-    GOOGLE,
-    ;
-}
